backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
bcrypt==4.1.2
passlib==1.7.4
